Add configurations to build binary wheels on GNU/Linux and Windows
Thanks to cibuildwheel, now every commit will trigger a build of manylinux1
wheels for Python 3.4, 3.5 and 3.6, both on GNU/Linux (travis-ci.org) and
Windows (ci.appveyor.com). Each build will be followed by a run of the test
suite.

On tags, if everything is ok, the wheel will be eventually automatically
uploaded to PyPI.

This fixes issue #67.