Handle generic iterables
This targets issue #84, automatically dumping generic iterables as arrays.

To overcome the inefficiency in handling generic iterables with the previous stack-based
implementation of the dumps_internal() function, this replaces it with plain recursion so
there's no need to build a temporary list consuming the iterable and then pushing each
item in reverse order.