Introduce ability to use already serialized objects
When the `default` function returns a bytes object, include it
verbatim into the serializaion.
This allows users to store preserialized versions of their objects in a database
and use rapidjson to generate structures that include them.
If the `default` function returns an invalid JSON the resulting string returned by
rapidjson will be in turn invalid.