Be sure to free copied keys when the PyHandler is destroyed
Investigating on issue #117, I found that in some way it's possible that the contexts
stack is not empty at the end of the job, when the PyHandler gets destroyed: some still
alive context could be pointing to memory allocated for dictionary keys, when parsing in
non-insitu mode. Since that currently happens only when loading from a stream, it could be
very well (one of) the cause(s) for the reported memory leak. Unfortunately the OP was not
able to confirm.