Use a std::string to hold the dictionary keys to be sorted
For some reason, certainly beyond my early morning understanding of C++
behaviour, the call to std::sort() on the std::vector<DictItem>
scrambles the items data, and indeed the array after the call contains
items matching the effective output shown in the issue #229, that is,
the first two items have the same key, "2025-11-11".

This replaces low level C string, a pointer to char plus the length,
with a std::string value, to hold the dictionary keys being sorted,
and apparently this cures the problem.